fix detekt issues
