+++ app/src/main/java/ru/practicum/android/diploma/chooseplacework/ui/ChooseWorkPlaceFragment.kt
@@ -1,14 +1,13 @@
-package ru.practicum.android.diploma.choose_place_work.ui
+package ru.practicum.android.diploma.chooseplacework.ui
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import ru.practicum.android.diploma.R
 import ru.practicum.android.diploma.databinding.FragmentChooseWorkPlaceBinding
-import ru.practicum.android.diploma.databinding.FragmentFiltersBinding
 
 class ChooseWorkPlaceFragment : Fragment() {
 
@@ -16,7 +15,8 @@ class ChooseWorkPlaceFragment : Fragment() {
     private val binding get() = _binding!!
 
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
+        inflater: LayoutInflater,
+        container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
         _binding = FragmentChooseWorkPlaceBinding.inflate(inflater, container, false)
