+++ app/src/main/java/ru/practicum/android/diploma/chooseregion/ui/ChooseRegionFragment.kt
@@ -1,12 +1,10 @@
-package ru.practicum.android.diploma.choose_region.ui
+package ru.practicum.android.diploma.chooseregion.ui
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import ru.practicum.android.diploma.R
-import ru.practicum.android.diploma.databinding.FragmentChooseCountryBinding
+import androidx.fragment.app.Fragment
 import ru.practicum.android.diploma.databinding.FragmentChooseRegionBinding
 
 class ChooseRegionFragment : Fragment() {
@@ -15,7 +13,8 @@ class ChooseRegionFragment : Fragment() {
     private val binding get() = _binding!!
 
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
+        inflater: LayoutInflater,
+        container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
         _binding = FragmentChooseRegionBinding.inflate(inflater, container, false)
