+++ app/src/main/java/ru/practicum/android/diploma/core/ui/root/RootActivity.kt
@@ -11,12 +11,13 @@ import ru.practicum.android.diploma.databinding.ActivityRootBinding
 
 class RootActivity : AppCompatActivity() {
 
-    private lateinit var binding: ActivityRootBinding
+    private var _binding: ActivityRootBinding? = null
+    private val binding get() = _binding!!
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        binding = ActivityRootBinding.inflate(layoutInflater)
+        _binding = ActivityRootBinding.inflate(layoutInflater)
         setContentView(binding.root)
 
         val navHostFragment = supportFragmentManager.findFragmentById(R.id.rootFragmentContainerView) as NavHostFragment
@@ -31,9 +32,10 @@ class RootActivity : AppCompatActivity() {
                 R.id.chooseIndustryFragment,
                 R.id.chooseRegionFragment,
                 R.id.chooseWorkPlaceFragment,
-                R.id.filtersFragment-> {
+                R.id.filtersFragment -> {
                     binding.bottomNavigationView.visibility = View.GONE
                 }
+
                 else -> {
                     binding.bottomNavigationView.visibility = View.VISIBLE
                 }
