+++ app/src/main/java/ru/practicum/android/diploma/favorites/ui/FavoritesFragment.kt
@@ -1,14 +1,13 @@
 package ru.practicum.android.diploma.favorites.ui
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import ru.practicum.android.diploma.R
 import ru.practicum.android.diploma.databinding.FragmentFavoritesBinding
-import ru.practicum.android.diploma.databinding.FragmentVacancySearchBinding
 
 class FavoritesFragment : Fragment() {
 
@@ -16,7 +15,8 @@ class FavoritesFragment : Fragment() {
     private val binding get() = _binding!!
 
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
+        inflater: LayoutInflater,
+        container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
         _binding = FragmentFavoritesBinding.inflate(inflater, container, false)
