+++ app/src/main/java/ru/practicum/android/diploma/team/ui/TeamFragment.kt
@@ -1,12 +1,10 @@
 package ru.practicum.android.diploma.team.ui
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import ru.practicum.android.diploma.R
-import ru.practicum.android.diploma.databinding.FragmentFiltersBinding
+import androidx.fragment.app.Fragment
 import ru.practicum.android.diploma.databinding.FragmentTeamBinding
 
 class TeamFragment : Fragment() {
@@ -15,7 +13,8 @@ class TeamFragment : Fragment() {
     private val binding get() = _binding!!
 
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
+        inflater: LayoutInflater,
+        container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
         _binding = FragmentTeamBinding.inflate(inflater, container, false)
