+++ app/src/main/java/ru/practicum/android/diploma/vacancysearch/ui/VacancySearchFragment.kt
@@ -1,25 +1,25 @@
-package ru.practicum.android.diploma.settings_filter.ui
+package ru.practicum.android.diploma.vacancysearch.ui
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
 import ru.practicum.android.diploma.R
-import ru.practicum.android.diploma.databinding.FragmentFiltersBinding
 import ru.practicum.android.diploma.databinding.FragmentVacancySearchBinding
 
-class FiltersFragment : Fragment() {
+class VacancySearchFragment : Fragment() {
 
-    private var _binding: FragmentFiltersBinding? = null
+    private var _binding: FragmentVacancySearchBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
+        inflater: LayoutInflater,
+        container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentFiltersBinding.inflate(inflater, container, false)
+        _binding = FragmentVacancySearchBinding.inflate(inflater, container, false)
         return binding.root
     }
 
@@ -27,11 +27,11 @@ class FiltersFragment : Fragment() {
         super.onViewCreated(view, savedInstanceState)
 
         binding.button.setOnClickListener {
-            findNavController().navigate(R.id.action_filtersFragment_to_chooseIndustryFragment)
+            findNavController().navigate(R.id.action_vacancySearchFragment_to_vacancyFragment)
         }
 
         binding.button2.setOnClickListener {
-            findNavController().navigate(R.id.action_filtersFragment_to_chooseWorkPlaceFragment)
+            findNavController().navigate(R.id.action_vacancySearchFragment_to_filtersFragment)
         }
     }
