+++ app/src/main/res/layout/fragment_choose_industry.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".choose_industry.ui.ChooseIndustryFragment">
+    tools:context=".chooseindustry.ui.ChooseIndustryFragment">
 
     <TextView
         android:layout_width="match_parent"
