+++ app/src/main/res/layout/fragment_choose_work_place.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".choose_place_work.ui.ChooseWorkPlaceFragment">
+    tools:context=".chooseplacework.ui.ChooseWorkPlaceFragment">
 
     <TextView
         android:layout_width="match_parent"
