+++ app/src/main/res/layout/fragment_filters.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".settings_filter.ui.FiltersFragment">
+    tools:context=".settingsfilter.ui.FiltersFragment">
 
     <TextView
         android:layout_width="match_parent"
