+++ app/src/main/res/layout/fragment_vacancy_search.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".vacancy_search.ui.VacancySearchFragment">
+    tools:context=".vacancysearch.ui.VacancySearchFragment">
 
     <TextView
         android:layout_width="match_parent"
